+++ nodes/web/react@1.0.0/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nanoservice-ts/react
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.26
+
 ## 0.0.16
 
 ### Patch Changes