+++ triggers/grpc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nanoservice-ts/trigger-grpc
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.26
+  - @nanoservice-ts/if-else@0.0.30
+  - @nanoservice-ts/api-call@0.1.29
+
 ## 0.0.13
 
 ### Patch Changes